keystone: dts: disable 10G bindings by default
The EVM requires additional add on cards to support 10G. As the hardware
is not present in the EVM, by default it throws some error to the console
and is undesirable. Right thing to do is to keep the binding for 10G
disabled and enable it when the board has the required hardware to
support 10G Ethernet. This patch keeps it disabled by default. Some
suggested options to use for testing are

 - enable the bindings through u-boot boot command by adding following
   to the bootcmd.

   fdt addr ${addr_fdt};
   fdt set /soc/i2c2/pca@20 status "ok";
   fdt set /soc/netcp@2f00000 status "ok";
   fdt set /soc/mdiox/ status "ok";

Or

   update the status to "ok" in DTS for the above nodes and rebuild
   DTB and use. User may extract the existing DTS file

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>