net: keystone: Fix DMA unmap buffer sizes
This patch fixes the buffer sizes passed to dma_unmap_single() to
be the size previously mapped, rather than the size of the entire
allocation. This avoids unmapping areas that were not previously
mapped.

Also extended the #ifdef DEBUG around the packet dump functions,
since this only necessary when debugging.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>