crypto: keystone: Fix Command label buffer overflow
The maximum size of PSDATA allowed in a 128byte pktdma descriptor
is 64 bytes. Keystone crypto driver sends command-label & the
crypto request context identifier in PSDATA. The driver needs
to ensure that the total size of these don't exceed the size of
PSDATA. Also there was a bug in the driver code that the buffer
size for holding this data was not sufficient.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>