arm: arch_timer: increase arch_timer precision.
On systems with high arch_timer_rate, which is close to the NSEC_PER_SEC
value, the sched_clock_mult rounded to integer may have huge inaccuracy
and timer may fire significantly earlier than it should.

As far as the arch timer has only 56 bit, we may use additional 8 bits
as a fractional part for the arch_timer_rate. Actually in this patch we
use only 7 bits. Evan that increased accuracy in 128 times.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>