rapidio: keystone: fix kernel link warning (wrong section usage)
The KeyStone RapidIO driver was early started on a 2.6.34 kernel
where probe functions from platform_driver definition must be
defined with __devinit or __init qualifiers.
With recent kernel 3.x this has been removed for generic
hot-plug support purpose (probe are now standard static functions)
but the KeyStone RapidIO driver was not modified in that sense.
As a consequence a section mismatch warning was displayed during
the Linux kernel image link for the keystone_rio_probe() function.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>