net: keystone: Export CPSW port number as interface "dev_id"
This patch exports the CPSW port number as the interface "dev_id".

This field previously was always zero. If the NetCP instance is
configured in single-interface mode, it will remain zero. In
multi-interface mode, the number is the CPSW port number in the
range 1 to n, where "n" is the number of external ports (4 for
the k2hk, 8 for the k2e, etc).

Signed-off-by: Reece R. Pollack <x0183204@ti.com>