dts: keystone: Add NetCP PA PDSP firmware file name definitions
Prior to this patch, the firmware names used to load the NetCP PA
PDSPs was hard-coded. This is especially problematic for the older
PA, where one firmware image was loaded into multiple PDSPs. Changing
this should not require kernel source code.

This patch adds PA PDSP firmware names to the DTS in preparation
for a corresponding change to the NetCP PA driver.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>