net: keystone: Get NetCP PA PDSP firmware file names from DTS
Prior to this patch, the firmware names used to load the NetCP PA
PDSPs was hard-coded. This is especially problematic for the older
PA, where one firmware image was loaded into multiple PDSPs. Changing
this should not require kernel source code modifications.

This patch change the NetCP PA drivers to get the PDSP firmware
file names from the DTS.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>