spi: davinci: fix spidev ioctl failure
user space application uses spidev device (/dev/spidev.xxx) to
communicate with slave devices, for example FPGA and every time it
calls setup() API, gpio_request() call is made to setup GPIO CS if in
use. This fails for the second time and subsequent calls if GPIO is
not freed. Add a gpio_free() call to free the GPIO if already in use
to fix this. Also add a check for return value gpio_direction_output()
call and take action. In addition, call gpio_free() in
davinci_spi_cleanup() to free up the GPIO pin in use.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>