hwqueue: keystone: Change QoS WRR to allow 1:10000 ratio
This patch changes the algorithm used to convert WRR ratios
to PDSP credits to allow for wide (1:10000) ratios.

The previous algorithm attempted to keep the mean of the
WRR credits to be 3000. This works with narrow ratios, but
with wide ratios the lower-weight queues received very small
credit values causing excessive overhead and a significant
degradation in throughput.

The new algorithm attempts to assign a credit of 1500 bytes
to the lowest-weight queue, and calculates higher-weight
queue credits accordingly. However, if this would result in
an overflow, the lowest-weight queue credit is divided by
two iteratively until the highest-weight queue credit fits.
A warning is generated if this results in an excessively
small credit value.

This patch also makes use of the updated WRR shifts to
allow a wider range of credit values, and thus must be
used with firmware 2.0.1.8.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>