hwqueue: keystone: Fix WARN assertion when CONFIG_DEBUG_MUTEXES enabled
During QoS initialization, the pdsp clock rate is needed to configure
the timer rate. This is done by calling clk_get_rate() which attempts to
acquire a mutex, but a spinlock is already held. This lock sequence
violation results in the WARN assertion. This patch determines the PDSP
clock rate before the spinlock is acquired.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>