net: keystone: Handle unexpected RX interrupts more gracefully
On extremely rare occasions, RX interrupt notifications have been
observed with RX state other than INTERRUPT, which resulted in a
BUG assertion. This seems to occur only during interface shutdown,
even though RX interrupts should have ceased before the state is
set to TEARDOWN.

This patch specifically checks for an interrupt notification with
the state in TEARDOWN and recovers by immediately returning. If
the state is other than INTERRUPT or TEARDOWN, a WARN is asserted
rather than a BUG to allow the system to continue more gracefully.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>