phy: Add serdes phy driver for keystone
This phy driver added to support keystone PCI driver. This is
a generic phy driver and currently it supports only SerDes
for PCI Controller. The hw vendor that provides the phy
hw published only registers and their values. So this driver
uses these hard coded values to initialize the phy.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>