net: keystone: Don't re-initialize CPSW host port and ALE
Prior to this patch, every time an interface was brought up
the CPSW host port and a number of ALE configurations were re-
initialized. This is unnecessary and interferes with some uses
of the CPSW by DSPs. This patch causes the CPSW host port and
ALE to be configured only when the first interface on the CPSW
is brought up.

Note that if all of the interfaces on the CPSW are brought down,
the clocks driving the CPSW (and other parts of the NetCP) are
shut off for power consumption, which is assumed to cause a loss
of configuration. Thus the CPSW must be re-initialized when the
first interface is brought back up.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>