soc: ti: enhancement to support qmss on nss lite
NSS Lite has a cut down version of QMSS. It doesn't have status and
explicit push register space as in QMSS available on other K2 SoCs.
So definite reg indices specific to QMSS on NSS lite. This patch
introduces "keystone-navigator-qmss-l" compatibility to identify this
version of the hardware and to customize the dts handling code.

Update the Documentation for DT bindings to reflect this change

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>