net: keystone: properly set destination port in netcp 1.5
When transmitting packets in directed port mode, the port
number of the port on which the packet to be sent out needs
to be set properly in the packet descriptor.
In netcp 1.5 (K2L/E/G), such port number is set to the
destination tag field of the packet descriptor, while in
netcp 1.0 (K2HK) it is set to the psflags field.  Changes in
this patch make sure it is set properly.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>