net: keystone: update subsystem version NU statistics definitions
This patch updates/consolidates statistics definitions
for the ethernet switch module in netcp 1.5.

Notice that not all statistics registers are applicable to all
versions of the ethernet switch in netcp 1.5.  Some are for
NU version only and not available for 2U version.

Also host port statistics registers is a proper subset of those
of slave ports.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>