rapidio/rio_cm: add loopback device for local mport
Add loopback peer device to allow applications send/receive channalized
MBOX messagess using the same local mport device.

Currently this patch adds loopback capability for testing purposes only
and has to be used with caution:

The loopback initialization adds a local mport as peer to itself without
checking if SRIO link is functional and is capable to route packets between
TX and RX channles of the mport.

Normally this operation should be allowed only if there is an SRIO switch
connected to the mport lines. To enable testing using mport's internal
loopback capability (if available) check for correct HW configuration is
not implemented.

Signed-off-by: Alexandre Bounine <alexandre.bounine@idt.com>