usb: keystone: handle irq directly instead of through irq-chip
Currently the usb irq is handled through the implementation of
an irq-chip.  But since the irq is solely for usb, there is no
need to handle the irq through such an indirect route.  This patch
removes the irq-chip implementation and allow the driver handles
irq directly.  Direct handling should improve the performance
and have cleaner usb driver code.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>