usb: keystone: dts instantiation of platform specific phy and dwc3 core drivers
Current implementation registers and starts the generic
nop phy driver and dwc3 core driver from within the
dwc3 keystone specific glue driver.  This is not scalable
in the sense that supporting more than one usb controller
would be difficult.  This patch adds a keystone specific
phy driver.  Both of this phy driver and the dwc3 core driver
are instantiated through devicetree bindings.  This
approach makes it easier to implement the support of more
than one USB controller.

This is is the same approach taken in the upstreamed version
of the keystone usb driver.  The upstreamed version cannot be
backported cleanly due to API changes between kernel versions.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>