dma: keystone: add capability to specify packet type at prep
This patch allows the dma-engine user layer to specify the packet type on a per
packet basis at transfer prep time. The expectation is that the caller passes
in an "options" argument with the DMA_HAS_PKTTYPE flag set, and with the
packet type information included (see DMA_PKTTYPE_MASK and DMA_PKTTYPE_SHIFT).

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>

Conflicts:

	drivers/dma/keystone-pktdma.c
	include/linux/keystone-dma.h