keystone: dts: disable PCIe by default for K2L EVM
K2L EVM has by default SGMII configured to use the two serdes lanes and
the same is not available for PCIe. Currently both SGMII driver and
PCI driver is configured to use the SerDes. This patch disables PCIe-0
by default in keystone.dtsi and enable it only for K2HK and K2E. For K2L
it will remain disabled to match with hardware configuration. To enable
PCI-0, status needs to be set to "ok" on K2L based boards that has mux in
configured for PCIe to use SerDes lanes. For PCI-1, we don't have the
support enabled in DTS. So no issue.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>