rapidio: keystone: drop address alignment in LSU management
The RapidIO load-store unit (LSU) within Keystone II processors handles
non-aligned addresses correctly. The destination address and byte count passed
to the LSU is used to setup a correct RapidIO packet header with 64 bit
alignment, even if the original destination address isn't 64 bit aligned.

As this support is present in the LSU for NREAD, NWRITE and NWRITE_R, it
isn't necessary to enforce 64 bit alignment of addresses at the driver level.

But because SWRITE requires a strict alignment on 64 bit, we only switch to
SWRITE mode when both source and destination addresses are both aligned on
64 bit.

Signed-off-by: Paul Barker <paul.barker@commagility.com>
Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>