hwqueue: Fix race in disabling interrupt notifications
Prior to this patch, there existed a race condition between
disabling of notifications in hwqueue_disable_notifier() and
hwqueue_notify(). In this race, the delivery of an interrupt
after the notifier was marked disabled but before disabling
the interrupt would result in the interrupt service routine
not calling the notifier function. Thus the interrupt persists
which prevents or severely delays the disabling of the interrupt.

This patch synchronizes the entire notifier disable sequence
against the delivery of an interrupt notification by the use
of a spinlock. Synchronization is also provided against the
enable sequence. Note that the use of atomic operations does
not suffice in this situation.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>