net: Fix skb->protocol setting for mixed IPsec tunnels
When processing fragmented packets in a mixed-protocol tunnel
(i.e. IPv6 inner and IPv4 outer, or IPv4 inner and IPv6 outer),
the skb->protocol field is being set to the wrong value. This
patch sets the skb->protocol value according to the IP header
version of the packet it's carrying.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>