net: keystone: Remove erroneous dmaengine_pause calls in PA shutdown
This patch removes erroneous calls to dmaengine_pause() in the NetCP
PA shutdown sequence. This function should be called to stop DMA
completion notifications that would trigger later dma_poll() calls
from a softIRQ or tasklet. If the interrupt is to be serviced from
the notifier routine, then we should not be pausing the channel
even if we're shutting it down.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>