dma: keystone: Fix memory leak freeing DMA descriptors
This patch corrects a memory leak when freeing software DMA
descriptors. The loop containing the error was intended to
identify any hardware DMA descriptors which did not get freed
by the previous loop, but do to missing braces ({..}) it always
set "leaked" to "true" which prevented the software descriptor
array from being freed.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>