net: keystone: Fix memory leak in CPSW ALE configuration
Prior to this patch, the CPSW ALE context was created when the first
interface on a device was opened, but it was not destroyed when the
last interface was closed (it was stopped instead). This resulted in
a memory leak. This patch corrects this.

Also, the ALE Slaves were being shut down after the ALE context was
being destroyed. This could result in a NULL pointer reference. This
patch fixes this too.

Finally, the ALE reference count in keystone_xgess.c was a non-atomic
integer, while the other modules used an atomic. This patch fixes this
and includes some minor restructuring so the code looks more like the
other two modules.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>