net: keystone: Remove NetCP CPSW ALE VLAN configuration
An early implementation of the NetCP CPSW module configured the
Address Lookup Engine (ALE) to be VLAN-aware. This was later
determined to be undesirable, and the VLAN-aware bit in the ALE
Control register was changed to 0. However, the code to configure
VLAN-specific entries remained. This patch removes this unneeded
code, and addresses a NULL pointer dereference in the process.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>