net: keystone: clean up slave port configurations
In multi-if mode:

When MDIO/PHY is enabled, carrier on/off is reported
to upper layer by the phy state machine, hence cpsw
driver needs to report carrier on/off when the link
interface type is other than MAC-to-PHY.

Also, in the MAC-to-PHY case, in addition to checking
the sgmii link status register, cpsw driver needs to
check the bit that corresponds to the phy connected to
the slave port in the phy link state variable in netcp
in order to conclude the sgmii link status.  Only when
both of these two statuses show an up state can the cpsw
driver conclude the link is up.  Before this
patch, it checks the aggregated value of the phy link
state variable.

In the case of MAC-to-MAC, the cpsw driver is responsible
to report the carrier on/off status to upper layers.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>