net: keystone: fix corruption/hangs after concurrent update of statistics
When there are multiple writers to the stats, such as when a frame is
transmitted and a frame is received at the same time on different CPU,
The sequence counter used to synchronize access to the network statistics
can become corrupted. This can result in one of the CPUs spinning forever.
This patch fixes this by separating the TX and RX statistics, each protected
by its own sequence counter.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>