net: keystone: Fix a bug in an error recovery path
If dma_request_channel_by_name() returns an error, the return value
may be a non-NULL pointer. The fail path then tries to release the
channel using the bogus pointer, which may cause an oops or crash.

This patch corrects this bug by explicitly setting the channel pointer
to NULL before jumping to the fail path.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>