hwqueue: keystone: Don't set output throttle on collapsed ports
When setting a new output rate on a QoS port we also need to set
the output throttle threshold of the ports that feed that port.
However, when the input is from drop schedulers what appears to
be a "port" in the DTS may be optimized away. In this case,
attempting to set the child port's throttle threshold is actually
setting the target port instead. This patch detects this condition
and prevents the throttle threshold from being set.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>