net: keystone: Change NetCP QoS queue mapping table format
Prior to this patch, the NetCP QoS queue mapping table was interpreted
as an ordered sequence of entries, with the first entry representing
a skb->queue_mapping value of 0, the next of 1, etc. However, this
implicit ordering is not preserved by the DTB decompiler (which reverses
the order of the table), and the results of merging multiple definitions
is not predictable.

The convention has been to name the table entries "chan-0", "chan-1", etc.,
even though this has not been enforced. This patch enforces this existing
convention, and uses the number suffix to represent the index into the
queue_mapping table.

Note that this now allows for sparse mapping tables, should that be desired.
Packets that arrive at the interface with invalid skb->queue_mapping values
will not be mapped through the table, leaving them to be handled by whatever
TX hook previously claimed the packet.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>