net: keystone: consolidate gbe and 10gbe serdes configurations
This patch consolidates the gbe and 10gbe serdes configurations
into one common framework.  This is a step towards the unification
of serdes configurations of Keystone2 peripherals that use serdes.
In particular, it adds support of per serdes and per lane configurations.
Among the Keystone2 peripherals that use serdes are gbe, 10gbe and pcie.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Michael Scherban <m-scherban@ti.com>