rapidio: keystone: add option for error recovery and use port reset instead
This driver module option allows to disable the error recovery mechanism
initiated in case of port error for aligning ackIds.
But this error recovery mechanism may not work with all kind of link
partners such as RapidIO CPS switches or some FPGA end-points.
In that case, error recovery can be disabled with this option. Instead
a reset symbol is sent to the corresponding port and the whole RapidIO
controller is reset.

By default error recovery is enabled in the driver. Use the
'keystone_rio_mport.error_recovery=0' kernel command line option to
disable it.

Signed-off-by: Aurelien Jacquiot <a-jacquiot@ti.com>