net: keystone: fix 'opened' variable bug
The 'opened' variables are shared across devices that can have multiple
interfaces and is meant to indicate when the device is opened or closed.

This fixes a bug where closing one interface on the device sets 'opened'
to 0 when other interfaces may still be alive. Instead, the reference
and interface counters will be used to track the status of the device.

Signed-off-by: Michael Scherban <m-scherban@ti.com>