net: keystone: init PA firmware during probe
Currently the PA is restarted and reconfigured when the number of
network interfaces goes from 0 to 1. This is not ideal under some
configurations where the PA is a shared resource outside of the kernel.

With this patch the PA firmware will be configured once during boot,
rather than during runtime, so that any other systems using the PA will
continue to run.

Signed-off-by: Michael Scherban <m-scherban@ti.com>