net: keystone: cable unplug lockup workaround for K2E and K2L gbe
On K2E/L gbe, if the ethernet cable is unplugged when the switch is
sending packets to the network, it can sometimes causes a system lockup.

The issue was created when the GMII_EN bit is cleared and a packet was
being transmitted. This resets the MAC state machine.  When the GMII_EN
is then set back to a ‘1’, the MAC state machine detects that the FIFO
is not at start of frame and proceeds to purge the rest of the frame.

When the end of frame is detected, it issues a tx_stat_valid, but since
the MAC state machine was not in the transmit states, there is no status
(good, collisions , etc).  When the stat block sees the tx_stat_valid it
goes to the transmit state and since there is no status there was no exit
condition. The next time stats are read will hang the bus.

This patch provides a workaround that avoids the lockup by disabling the
port statistics when the link goes down.  When the link comes back up, the
MAC is reset and port statistics is re-enabled when the MAC becomes
idle after reset.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>