hwqueue: keystone: fix race on concurrent accumulator push and pop
When popping from an accumulator queue we are popping from an array
instead of a hardware queue. On the push side the driver increments a
variable, desc_count, when a descriptor is available and pushes the
address to the array. On the pop side the array will be read if
desc_count is incremented.

On setups that poll the accumulator it has been observed that a race can
occur when desc_count is incremented and the read happens before the
write; causing an old descriptor address to be read.

This prevents the concurrent access to the data by spinlocking for the
increment of desc_count and the write to the array during the push, and
spinlocking before desc_count is read on the pop.

Signed-off-by: Michael Scherban <m-scherban@ti.com>