net: keystone: Use RCU-bh to reduce rcu_preempt stall reports
Under certain heavy load conditions, the -RT kernels report rcu_preempt
self-detected stalls. The only explicit use of RCU in the NetCP driver
is to protect the TX and RX hook lists, which are almost entirely read-
only; they're only updated when an interface is brought up or down, or
a module is inserted or removed while an interface is up.

The dispatch code in net/core/dev.c uses the -bh flavor of RCU to avoid
stalls when under a DDOS attack, which is the situation the customer
reporting the problem is testing. Thus it makes sense to change NetCP's
use to the -bh flavor of RCU as well.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>